tests/console/console-clear-expected.txt

CONSOLE MESSAGE: line 9: one
CONSOLE MESSAGE: line 10: two
CONSOLE MESSAGE: line 11: three
Tests that console is cleared upon requestClearMessages call.

=== Before clear ===
console-clear.html:9 one
console-clear.html:10 two
console-clear.html:11 three
=== After clear ===
